+++ hazelcast/src/main/java/com/hazelcast/map/impl/operation/PutAllBackupOperation.java
@@ -21,8 +21,8 @@ import com.hazelcast.internal.serialization.Data;
 import com.hazelcast.map.impl.MapDataSerializerHook;
 import com.hazelcast.map.impl.record.Record;
 import com.hazelcast.map.impl.record.Records;
-import com.hazelcast.map.impl.recordstore.ExpiryMetadata;
-import com.hazelcast.map.impl.recordstore.ExpirySystem;
+import com.hazelcast.map.impl.recordstore.expiry.ExpiryMetadata;
+import com.hazelcast.map.impl.recordstore.expiry.ExpiryMetadataImpl;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.spi.impl.operationservice.BackupOperation;
@@ -121,7 +121,7 @@ public class PutAllBackupOperation extends MapOperation
         for (int i = 0; i < size; i++) {
             keyRecordExpiry.add(IOUtil.readData(in));
 
-            ExpiryMetadata expiryMetadata = new ExpirySystem.ExpiryMetadataImpl();
+            ExpiryMetadata expiryMetadata = new ExpiryMetadataImpl();
             keyRecordExpiry.add(Records.readRecord(in, expiryMetadata));
             keyRecordExpiry.add(expiryMetadata);
         }
