+++ hazelcast/src/main/java/com/hazelcast/map/impl/recordstore/AbstractEvictableRecordStore.java
@@ -26,6 +26,8 @@ import com.hazelcast.map.impl.MapContainer;
 import com.hazelcast.map.impl.event.MapEventPublisher;
 import com.hazelcast.map.impl.eviction.Evictor;
 import com.hazelcast.map.impl.record.Record;
+import com.hazelcast.map.impl.recordstore.expiry.ExpiryReason;
+import com.hazelcast.map.impl.recordstore.expiry.ExpirySystem;
 import com.hazelcast.spi.impl.NodeEngine;
 import com.hazelcast.spi.impl.eventservice.EventService;
 import com.hazelcast.spi.merge.SplitBrainMergeTypes.MapMergeTypes;
@@ -105,11 +107,6 @@ public abstract class AbstractEvictableRecordStore extends AbstractRecordStore {
     public boolean shouldEvict() {
         Evictor evictor = mapContainer.getEvictor();
         return evictor != NULL_EVICTOR && evictor.checkEvictable(this);
-    }
-
-    @Override
-    public boolean isTtlOrMaxIdleDefined(Record record) {
-        return expirySystem.isTtlOrMaxIdleDefined(record);
     }
 
     @Override
